requirements.txt
tensorflow-gpu>=1.5.0
opencv-python
scipy
scikit-image
h5py
imageio-ffmpeg
pandas
tqdm
requests
ray